feat: add multiple integer types (i8, i16, i32, u8, u16, u32, u64)
Add IntWidth enum with 8 variants and change Type::Int to
Type::Int(IntWidth). Integer literals are now polymorphic — they get a
fresh type variable constrained to integer types, with the concrete
type determined by context and defaulting to i64. Cross-type arithmetic
(e.g. i32 + i64) is a type error. Unsigned negation is rejected.

Wasm codegen maps i64/u64 to ValType::I64 and all smaller types to
ValType::I32 with appropriate truncation/sign-extension. Component ABI
store functions are filled in for all widths. Overflow checking is
generalized per type.

+++ starstream-compiler/src/parser/primitives.rs
@@ -26,7 +26,7 @@ pub fn identifier<'a>() -> impl Parser<'a, &'a str, Identifier, Extra<'a>> {
 pub fn integer_literal<'a>() -> impl Parser<'a, &'a str, Literal, Extra<'a>> + Clone {
     text::int(10)
         .map(|digits: &str| {
-            let value = digits.parse::<i64>().expect("integer literal");
+            let value = digits.parse::<i128>().expect("integer literal");
             Literal::Integer(value)
         })
         .boxed()
